+++ GeoGen.AnalyticalGeometry/Constructors/Randomness/IRandomnessProvider.cs
@@ -10,24 +10,24 @@ namespace GeoGen.AnalyticalGeometry
     public interface IRandomnessProvider
     {
         /// <summary>
-        /// Generates a random decimal number that is at least 0 and less than 1.
+        /// Generates a random double number that is at least 0 and less than 1.
         /// </summary>
-        /// <returns>A random decimal number from the interval [0,1).</returns>
-        decimal NextDecimal();
+        /// <returns>A random double number from the interval [0,1).</returns>
+        double NextDouble();
 
         /// <summary>
-        /// Generates a random decimal number that is at least a given lower bound and less than a given upper bound.
+        /// Generates a random double number that is at least a given lower bound and less than a given upper bound.
         /// </summary>
         /// <param name="minValue">The lower bound.</param>
         /// <param name="maxValue">The upper bound.</param>
-        /// <returns>A random decimal number from the interval [minValue, maxValue).</returns>
-        decimal NextDecimal(decimal minValue, decimal maxValue);
+        /// <returns>A random double number from the interval [minValue, maxValue).</returns>
+        double NextDouble(double minValue, double maxValue);
 
         /// <summary>
-        /// Generates a random decimal number that is at least 0 and less than a given upper bound.
+        /// Generates a random double number that is at least 0 and less than a given upper bound.
         /// </summary>
         /// <param name="maxValue">The upper bound.</param>
-        /// <returns>A random decimal number from the interval [0, maxValue).</returns>
-        decimal NextDecimal(decimal maxValue);
+        /// <returns>A random double number from the interval [0, maxValue).</returns>
+        double NextDouble(double maxValue);
     }
 }
